Don't assume pod interface is first / single
In the future we may have multiple interface definitions, with and without pod
interface; with and without any interfaces; with pod interface first in the
list or in any other position.

Instead of assuming the pod interface is the first in the list, this patch
makes an effort to find the interface in the list based on its well known
attributes (type, bridge name).

Note that the implementation is not robust enough, and we probably need a
better way to match between API and domain interfaces.