Support multiple SR-IOV networks
This patch adopts changes from the latest SR-IOV device plugin master [1] that
allow us to attach a VMI to multiple SR-IOV Multus networks. (Before the
change, devices allocated for multiple networks could end up in different boot
order than configured.)

Among other things, this patch adds read access to network attachment
definitions to virt-controller role. This is to allow the controller to fetch
network CRD objects and determine their respective resource names. For this
same reason, it extends KubevirtClient object to include NetworkClient().

We do *not* maintain backwards compatibility with older SR-IOV device plugin
releases. This is as intended because SR-IOV support is still considered
experimental, and there is no good reason for us to support the device plugin
release that does not support multiple networks.

Updated documentation to reflect the fact that kubevirt is now compatible with
device plugin master (and *not* compatible with v1.0).

todo: get rid of vendor/* change

[1] https://github.com/intel/sriov-network-device-plugin/pull/26