Fix extracting network CRD definition from a different namespace
Despite what api schema suggests, we actually support attaching VMIs
to networks from namespaces that are different from VMI's. But the
code that was fetching CRDs and extracting resourceNames from their
annotations to implement SR-IOV bindings was not accommodating for
that, making virt-controller fail to create virt-launcher pod.

The CRD standard document is explicit that whenever a full form of
network name (<namespace>/<networkName>) is used in networks
annotation, then we should fetch from the specified namespace and not
the pod's one.

To make sure we no longer regress on support for explicitly namespaced
network attachments, this patch adds a functional test that attaches a
VMI to a network from a different namespace and confirms the VMI
boots.