Replace phoracek's network crd client with official client (forked)
This gets us closer to the official client. We are still using a fork because
the official client misses a feature of fake clientset to register network crd
objects for unit tests.

To avoid api conflicts between base client modules, we are pulling the
release-1.11 branch of the (forked) client.

The change includes unrelated changes in vendor tree that were triggered by
'make deps-update'.